# docs/api/content-request-module.openapi.yaml
openapi: 3.0.3
info:
  title: Bookmark Content Request Module API
  version: 1.2.0
  description: |
    The HTTP contract a **content request module** must implement so Bookmark
    can search an external catalog, start downloads for approved requests, and
    track their progress.

    **Your module is the server described by this document; Bookmark is the
    client.** Bookmark is pointed at the module via the `TRACKER_CLIENT_URL`
    and `TRACKER_CLIENT_API_KEY` environment variables and calls the paths
    below relative to that base URL.

    Every request except `GET /health` carries an `X-API-Key` header that the
    module must validate. `GET /health` is probed without credentials and must
    be unauthenticated.

    Any non-2xx response is treated as a hard failure (no retries): Bookmark
    logs the response body and surfaces the status code to its user. A module
    that is unreachable is reported as 503.

    See `docs/content-request-modules.md` in the Bookmark repository for the
    full developer guide, including the import-matching invariant that links
    completed downloads back to requests by folder name.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Example module base URL (whatever TRACKER_CLIENT_URL points to)
security:
  - ApiKeyAuth: []
...
paths:
...
  /languages:
    get:
      operationId: getLanguages
      tags: [Search]
      summary: Language taxonomy for the search filter
      description: |
        Returns the languages the module can filter searches by. The `id`
        values are the module's own taxonomy — Bookmark never interprets
        them, it only passes the user's selection back in the `languages`
        field of `POST /search`.

        Return an empty list if the catalog has no language concept; Bookmark
        then hides its language filter. A missing endpoint (404) is treated
        the same as an empty list.
      responses:
        "200":
          description: Available languages, in any order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LanguagesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /search:
    post:
      operationId: search
      tags: [Search]
      summary: Search the catalog
      description: |
        Called when a Bookmark user searches the request page. Return results
        **pre-parsed and clean** — resolve upstream formatting quirks
        (author/narrator splitting, series extraction, HTML stripping) inside
        the module. Bookmark displays the returned fields as-is.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results for the requested page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Shared secret configured as TRACKER_CLIENT_API_KEY in Bookmark. Sent
        on every request except GET /health.
...
  responses:
    Unauthorized:
      description: Missing or invalid X-API-Key header.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    SearchRequest:
      type: object
      required: [query, categories]
      properties:
        query:
          type: string
          description: Free-text search query entered by the user.
          example: project hail mary
        categories:
          type: array
          description: >-
            Content types to include. Always sent by Bookmark (defaults to
            all types when the user does not filter). A module whose catalog
            lacks a type simply returns no results for it.
          items:
            type: string
            enum: [audiobook, ebook, comics]
          example: [audiobook, ebook, comics]
        searchIn:
          type: array
          description: >-
            Fields to match against. Omitted entirely when the user does not
            restrict the search — search all fields in that case.
          items:
            type: string
            enum: [title, author, narrator, series, tags, description]
        languages:
          type: array
          description: >-
            Language IDs from the module's GET /languages taxonomy, passed
            through opaquely from the UI. Omitted when not filtering by
            language.
          items:
            type: integer
        perPage:
          type: integer
          description: >-
            Page size. Bookmark's UI currently sends 10, 25, 50 or 100, but
            modules should accept any positive integer. Pick a sensible
            default when omitted.
          minimum: 1
          example: 25
        offset:
          type: integer
          description: Zero-based result offset for pagination.
          minimum: 0
          default: 0
      additionalProperties: false
    Language:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          description: Language ID in the module's own taxonomy.
          example: 1
        name:
          type: string
          description: Display name shown in Bookmark's filter UI.
          example: English
    LanguagesResponse:
      type: object
      required: [languages]
      properties:
        languages:
          type: array
          description: Empty when the catalog has no language concept.
          items:
            $ref: "#/components/schemas/Language"
    SearchResponse:
      type: object
      required: [results, total]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        total:
          type: integer
          description: Total number of matches across all pages.
          minimum: 0
          example: 87
    SearchResult:
      type: object
      required: [id, title, contentType, categoryId]
      properties:
        id:
          type: integer
          description: >-
            Stable unique catalog item ID. Bookmark stores it and later calls
            POST /download/{id} and GET /image/{id} with it.
          example: 123456
        title:
          type: string
          example: Project Hail Mary
        author:
          type: string
          nullable: true
          example: Andy Weir
        narrator:
          type: string
          nullable: true
          example: Ray Porter
        series:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/SeriesEntry"
        description:
          type: string
          nullable: true
          description: Plain-text description (strip upstream HTML).
        contentType:
          type: string
          description: >-
            The semantic content type Bookmark routes on. The module owns the
            mapping from its upstream taxonomy to these values (e.g. comics
            filed under an ebook category upstream must still be reported as
            "comics").
          enum: [audiobook, ebook, comics]
        categoryId:
          type: integer
          description: >-
            Category ID in the module's own taxonomy. Informational only -
            Bookmark stores and displays it but never interprets it.
          example: 42
        categoryName:
          type: string
          description: Human-readable category label shown in the UI.
          example: Audiobooks - Sci-Fi
        size:
          type: string
          description: Human-readable size shown in the UI.
          example: 1.2 GiB
        language:
          type: string
          example: English
        fileType:
          type: string
          example: M4B
        tags:
          type: array
          items:
            type: string
        addedDate:
          type: string
          description: When the item was added upstream. ISO 8601 recommended.
          example: "2026-01-15T10:30:00Z"
    SeriesEntry:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: The Expanse
        number:
          type: string
          nullable: true
          description: Position within the series, as a string ("1", "2.5", …).
          example: "3"
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: >-
            Human-readable error. Bookmark logs the raw response body on any
            non-2xx status; the exact shape is up to the module.
          example: Unknown torrent hash